+++ message_sender/src/app/config/__init__.py
@@ -1,9 +1,8 @@
 # Local
-
-# Local
+from .config import CONTACT_CSV, CONTACT_CSV_TEST, PDF_FILE_TEST, SPECIAL_MESSAGE
 from .logger import create_logger
 
 
 LOGGER = create_logger()
 
-__all__ = []
+__all__ = ["CONTACT_CSV", "SPECIAL_MESSAGE", "CONTACT_CSV_TEST", "PDF_FILE_TEST"]